refactor: PR#11レビュー推奨改善を適用する
- TerminalWindow.tsx: `shark:` を `sharks:` に修正し `--sharks` フラグと統一
- SPECIES_COLORS の型を `Record<BoidSpecies, \`#\${string}\`>` に強化
- SPECIES_WEIGHT_TABLE・getRandomSpecies() を speciesUtils.ts に分離
- SPECIES_TOTAL_WEIGHT をモジュールトップレベルで事前計算
- SPECIES_WEIGHT_TABLE に `as const satisfies` を付与してリテラル型に昇格
- フォールバックコメントに浮動小数点精度誤差の説明を補記
- boidRenderer.ts の内側ループで rowData を先頭束縛しアクセスを最適化
